feat: rename tabs
